+++ Libplanet.Tests/Blockchain/Renderers/AnonymousRendererTest.cs
@@ -1,5 +1,3 @@
-using System;
-using Libplanet.Action;
 using Libplanet.Blockchain.Renderers;
 using Libplanet.Blocks;
 using Libplanet.Tests.Common.Action;
@@ -9,134 +7,12 @@ namespace Libplanet.Tests.Blockchain.Renderers
 {
     public class AnonymousRendererTest
     {
-        private static IAction _action = new DumbAction();
-
-        private static IAccountStateDelta _stateDelta =
-            new AccountStateDeltaImpl(_ => null, (_, __) => default, default);
-
-        private static IActionContext _actionContext =
-            new ActionContext(default, default, default, _stateDelta, default);
-
-        private static Exception _exception = new Exception();
-
         private static Block<DumbAction> _genesis =
             TestUtils.MineGenesis<DumbAction>(default(Address));
 
         private static Block<DumbAction> _blockA = TestUtils.MineNext(_genesis);
 
         private static Block<DumbAction> _blockB = TestUtils.MineNext(_genesis);
-
-        [Fact]
-        public void ActionRenderer()
-        {
-            (IAction, IActionContext, IAccountStateDelta)? record = null;
-            var renderer = new AnonymousRenderer<DumbAction>
-            {
-                ActionRenderer = (action, context, nextStates) =>
-                    record = (action, context, nextStates),
-            };
-
-            renderer.UnrenderAction(_action, _actionContext, _stateDelta);
-            Assert.Null(record);
-            renderer.RenderActionError(_action, _actionContext, _exception);
-            Assert.Null(record);
-            renderer.UnrenderActionError(_action, _actionContext, _exception);
-            Assert.Null(record);
-            renderer.RenderBlock(_genesis, _blockA);
-            Assert.Null(record);
-            renderer.RenderReorg(_blockA, _blockB, _genesis);
-            Assert.Null(record);
-
-            renderer.RenderAction(_action, _actionContext, _stateDelta);
-            Assert.NotNull(record);
-            Assert.Same(_action, record?.Item1);
-            Assert.Same(_actionContext, record?.Item2);
-            Assert.Same(_stateDelta, record?.Item3);
-        }
-
-        [Fact]
-        public void ActionUnrenderer()
-        {
-            (IAction, IActionContext, IAccountStateDelta)? record = null;
-            var renderer = new AnonymousRenderer<DumbAction>
-            {
-                ActionUnrenderer = (action, context, nextStates) =>
-                    record = (action, context, nextStates),
-            };
-
-            renderer.RenderAction(_action, _actionContext, _stateDelta);
-            Assert.Null(record);
-            renderer.RenderActionError(_action, _actionContext, _exception);
-            Assert.Null(record);
-            renderer.UnrenderActionError(_action, _actionContext, _exception);
-            Assert.Null(record);
-            renderer.RenderBlock(_genesis, _blockA);
-            Assert.Null(record);
-            renderer.RenderReorg(_blockA, _blockB, _genesis);
-            Assert.Null(record);
-
-            renderer.UnrenderAction(_action, _actionContext, _stateDelta);
-            Assert.NotNull(record);
-            Assert.Same(_action, record?.Item1);
-            Assert.Same(_actionContext, record?.Item2);
-            Assert.Same(_stateDelta, record?.Item3);
-        }
-
-        [Fact]
-        public void ActionErrorRenderer()
-        {
-            (IAction, IActionContext, Exception)? record = null;
-            var renderer = new AnonymousRenderer<DumbAction>
-            {
-                ActionErrorRenderer = (action, context, exception) =>
-                    record = (action, context, exception),
-            };
-
-            renderer.RenderAction(_action, _actionContext, _stateDelta);
-            Assert.Null(record);
-            renderer.UnrenderAction(_action, _actionContext, _stateDelta);
-            Assert.Null(record);
-            renderer.UnrenderActionError(_action, _actionContext, _exception);
-            Assert.Null(record);
-            renderer.RenderBlock(_genesis, _blockA);
-            Assert.Null(record);
-            renderer.RenderReorg(_blockA, _blockB, _genesis);
-            Assert.Null(record);
-
-            renderer.RenderActionError(_action, _actionContext, _exception);
-            Assert.NotNull(record);
-            Assert.Same(_action, record?.Item1);
-            Assert.Same(_actionContext, record?.Item2);
-            Assert.Same(_exception, record?.Item3);
-        }
-
-        [Fact]
-        public void ActionErrorUnrenderer()
-        {
-            (IAction, IActionContext, Exception)? record = null;
-            var renderer = new AnonymousRenderer<DumbAction>
-            {
-                ActionErrorUnrenderer = (action, context, exception) =>
-                    record = (action, context, exception),
-            };
-
-            renderer.RenderAction(_action, _actionContext, _stateDelta);
-            Assert.Null(record);
-            renderer.UnrenderAction(_action, _actionContext, _stateDelta);
-            Assert.Null(record);
-            renderer.RenderActionError(_action, _actionContext, _exception);
-            Assert.Null(record);
-            renderer.RenderBlock(_genesis, _blockA);
-            Assert.Null(record);
-            renderer.RenderReorg(_blockA, _blockB, _genesis);
-            Assert.Null(record);
-
-            renderer.UnrenderActionError(_action, _actionContext, _exception);
-            Assert.NotNull(record);
-            Assert.Same(_action, record?.Item1);
-            Assert.Same(_actionContext, record?.Item2);
-            Assert.Same(_exception, record?.Item3);
-        }
 
         [Fact]
         public void BlockRenderer()
@@ -147,14 +23,6 @@ namespace Libplanet.Tests.Blockchain.Renderers
                 BlockRenderer = (oldTip, newTip) => record = (oldTip, newTip),
             };
 
-            renderer.RenderAction(_action, _actionContext, _stateDelta);
-            Assert.Null(record);
-            renderer.UnrenderAction(_action, _actionContext, _stateDelta);
-            Assert.Null(record);
-            renderer.RenderActionError(_action, _actionContext, _exception);
-            Assert.Null(record);
-            renderer.UnrenderActionError(_action, _actionContext, _exception);
-            Assert.Null(record);
             renderer.RenderReorg(_blockA, _blockB, _genesis);
             Assert.Null(record);
 
@@ -173,14 +41,6 @@ namespace Libplanet.Tests.Blockchain.Renderers
                 ReorgRenderer = (oldTip, newTip, bp) => record = (oldTip, newTip, bp),
             };
 
-            renderer.RenderAction(_action, _actionContext, _stateDelta);
-            Assert.Null(record);
-            renderer.UnrenderAction(_action, _actionContext, _stateDelta);
-            Assert.Null(record);
-            renderer.RenderActionError(_action, _actionContext, _exception);
-            Assert.Null(record);
-            renderer.UnrenderActionError(_action, _actionContext, _exception);
-            Assert.Null(record);
             renderer.RenderBlock(_genesis, _blockA);
             Assert.Null(record);